frame_new_test

+++ net/sim/frame_2/load_trace.py
@@ -15,7 +15,7 @@ def load_trace(cooked_trace_folder=COOKED_TRACE_FOLDER):
         with open(file_path, 'rb') as f:
             for line in f:
                 parse = line.split()
-                cooked_time.append(float(parse[0]))
+                cooked_time.append(int(parse[0]))
         all_cooked_time.append(cooked_time)
         all_file_names.append(cooked_file)
